Refactor to  View - ViewModel- Model architecture.

+++ lib/repository/note_firebase_repository.dart
@@ -1,11 +1,16 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:sophia_hub/constant/firebase.dart';
-import 'package:sophia_hub/model/note.dart';
+import 'package:sophia_hub/model/activity.dart';
+import 'package:sophia_hub/model/note/note_image.dart';
+import 'package:sophia_hub/model/note/note_regular.dart';
+import 'package:sophia_hub/model/note/note.dart';
+import 'package:sophia_hub/model/note/note_type.dart';
 import 'package:sophia_hub/model/result_container.dart';
 import 'package:sophia_hub/repository/note_repository.dart';
+import 'package:sophia_hub/view/page/home/notes/image_note.dart';
 
-class NoteFirebaseRepository extends NoteRepository {
+class NoteFirebaseRepository extends NoteRepository<GenericNote> {
   late FirebaseFirestore _firestore;
 
   late CollectionReference notesRef;
@@ -32,30 +37,39 @@ class NoteFirebaseRepository extends NoteRepository {
       notesRef.orderBy(FirebaseKey.timeCreated, descending: true).limit(10);
 
   @override
-  Future<Result> create(Note note) async {
-    //TODO check internet
+  Future<Result> create(GenericNote note) async {
     try {
       if (!note.isValid()) {
         return Result(
             err: FormatException(
-          "Dữ liệu không hợp lệ",
-        ));
+              "Dữ liệu không hợp lệ",
+            ));
       }
 
       DocumentReference currentDoc = this.notesRef.doc();
-      // set emotion objects
-      CollectionReference activityRef =
-          currentDoc.collection(FirebaseKey.activities);
-      WriteBatch batch = _firestore.batch();
-      for (final emotion in note.activities) {
-        batch.set(activityRef.doc(emotion.id), emotion.toJson());
+
+      //Nếu là note bình thường
+      if (note is Note) {
+        // set emotion objects
+        CollectionReference activityRef =
+        currentDoc.collection(FirebaseKey.activities);
+        WriteBatch batch = _firestore.batch();
+        for (final emotion in note.activities) {
+          batch.set(activityRef.doc(emotion.id), emotion.toJson());
+        }
+        await batch.commit();
+
+        //Cập nhật danh sách lên server
+        await currentDoc.set(note.toJson());
+
+        // Cập nhật trường id
+
+      } else if (note is NoteImage) {
+        //TODO uploadImages here
+        // Step 01: upload image from uri
+        // Step 02: set image urls
       }
-      await batch.commit();
 
-      //Cập nhật danh sách lên server
-      await currentDoc.set(note.toJson());
-
-      // Cập nhật trường id
       note.id = currentDoc.id;
 
       return Result(data: note);
@@ -65,18 +79,21 @@ class NoteFirebaseRepository extends NoteRepository {
   }
 
   @override
-  Future<Result> delete(Note note) async {
+  Future<Result> delete(GenericNote note) async {
     try {
-      await notesRef.doc(note.id).delete().then((value) {
-        print("OKAY");
-      });
-      WriteBatch batch = this._firestore.batch();
-      for (final activity in note.activities)
-        batch.delete(notesRef
-            .doc(note.id)
-            .collection(FirebaseKey.activities)
-            .doc(activity.id));
-      await batch.commit();
+      await notesRef.doc(note.id).delete();
+
+      if (note is Note) {
+        WriteBatch batch = this._firestore.batch();
+        for (final activity in note.activities)
+          batch.delete(notesRef
+              .doc(note.id)
+              .collection(FirebaseKey.activities)
+              .doc(activity.id));
+        await batch.commit();
+      } else if (note is NoteImage) {
+        //TODO delete images here
+      }
 
       return Result(data: "Okay");
     } on Exception catch (e) {
@@ -97,46 +114,53 @@ class NoteFirebaseRepository extends NoteRepository {
   }
 
   @override
-  Future<Result<Note>> update(Note note) async {
+  Future<Result<GenericNote>> update(GenericNote note) async {
     try {
       //Cập nhật
-      await notesRef.doc(note.id).update(note.toJson());
 
-      // Xóa tất cả activity hiện tại trong note
-      final snapshot =
-          await notesRef.doc(note.id).collection(FirebaseKey.activities).get();
-      await Future.forEach(snapshot.docs, (QueryDocumentSnapshot e) async {
-        await notesRef
-            .doc(note.id)
-            .collection(FirebaseKey.activities)
-            .doc(e.id)
-            .delete();
-      });
+      if (note is Note) {
+        await notesRef.doc(note.id).update(note.toJson());
 
-      //Cập nhật activity mới
-      WriteBatch activitiesBatch = _firestore.batch();
-      for (final activity in note.activities) {
-        activitiesBatch.set(
-            notesRef
-                .doc(note.id)
-                .collection(FirebaseKey.activities)
-                .doc(activity.id),
-            activity.toJson());
+        // Xóa tất cả activity hiện tại trong note
+        final snapshot =
+        await notesRef.doc(note.id).collection(FirebaseKey.activities).get();
+        await Future.forEach(snapshot.docs, (QueryDocumentSnapshot e) async {
+          await notesRef
+              .doc(note.id)
+              .collection(FirebaseKey.activities)
+              .doc(e.id)
+              .delete();
+        });
+
+        //Cập nhật activity mới
+        WriteBatch activitiesBatch = _firestore.batch();
+        for (final activity in note.activities) {
+          activitiesBatch.set(
+              notesRef
+                  .doc(note.id)
+                  .collection(FirebaseKey.activities)
+                  .doc(activity.id),
+              activity.toJson());
+        }
+        await activitiesBatch.commit();
+      } else if (note is NoteImage) {
+        //TODO update NoteImage here
+
       }
-      await activitiesBatch.commit();
 
-      return Result<Note>(data: note);
+      return Result<GenericNote>(data: note);
+      return Result(data: Note());
     } on Exception catch (e) {
       return Result(err: e);
     }
   }
 
   @override
-  Future<Result<List<Note>>> loadMore() async {
-    List<Note> notes = [];
+  Future<Result<List<GenericNote>>> loadMore() async {
+    List<GenericNote> notes = [];
 
     try {
-      if(query == null) query = getNotesQuery();
+      if (query == null) query = getNotesQuery();
 
       if (lastDoc != null) query = query!.startAfterDocument(lastDoc!);
 
@@ -144,13 +168,32 @@ class NoteFirebaseRepository extends NoteRepository {
 
       //save last doc for pagination purpose
       if (docs.isNotEmpty) this.lastDoc = docs.last;
-      //TODO get activities
-      docs.forEach((e) {
-        Note quote = Note.fromJson(e.data() as Map<String, dynamic>)..id = e.id;
-        notes.add(quote);
+
+
+      await Future.forEach(docs, (QueryDocumentSnapshot doc) async {
+        if (doc.get("type") == NoteType.REGULAR.name.toLowerCase()) {
+          Note note = Note.fromJson(doc.data() as Map<String, dynamic>)
+            ..id = doc.id;
+
+            final listActivity =
+            await notesRef.doc(doc.id).collection(FirebaseKey.activities).get();
+            note.activities.addAll((listActivity).docs
+                .map((e) => Activity.fromJson(e.data())
+                ..icon =
+                    activities
+                        .firstWhere((element) => element.id == e.id)
+                        .icon));
+
+          notes.add(note);
+        } else if (doc.get("type") == NoteType.IMAGE.name.toLowerCase()) {
+          NoteImage noteImage = NoteImage()
+            ..id = doc.id;
+          //TODO
+        }
+
       });
 
-      return Result<List<Note>>(data: notes, err: null);
+      return Result<List<GenericNote>>(data: notes, err: null);
     } catch (e) {
       return Result(err: Exception(e), data: null);
     }
